+++ redis-starter/src/main/java/com/gomcarter/frameworks/redis/annotation/Lock.java
@@ -34,7 +34,7 @@ public @interface Lock {
      *
      * @return timeout
      */
-    long timeout() default 120L;
+    int timeout() default 120;
 
     /**
      * <p>
